Handle `childrenArePresentational`and change `includeText` default to true (#251)
# Overview

This commit makes a couple of changes to the `experimentalGetAccessibilityTree` function:

## `childrenArePresentational`

Some roles have a [`childrenArePresentational` attribute](https://www.w3.org/TR/wai-aria-1.1/#childrenArePresentational) set to true. That means:

> The DOM descendants are presentational. User agents SHOULD NOT expose descendants of this element through the platform accessibility API. If user agents do not hide the descendant nodes, some information may be read twice.

For us, that means we should not expose child elements as elements or text nodes within the main element. The text of these elements may still be used in the main element's accessible name. 

We did this by pulling node information from the [aria-query](https://github.com/A11yance/aria-query) library to determine the relevant roles. Then we check whether an element's role has the `childrenArePresentational` attribute set to true, before iterating over its children.

## `includeText` default value is now `true`

The change above reduced the amount of duplicate output when `includeText` is set to `true` so we made that the new default.